Updated Assembly File to Test Addition
Assembly files is updated to test addition, however I cannot properly recreate the ROM with the new instructions initialized, so I'm going to test on a new branch with upgrade RAM and ROM for the new version of Vivado. If I get no errors running it, I will merge it back to the master branch.